WIP: Refactored qp-klp plugin to support new pipelines
Refactored qp-klp plugin using mixins to break up tellseq.sh into
components and support a version 2.0 Tellread workflow in mg-scripts.

klp.py needs more cleanup and the tests need to be refactored. All
references to Step in klp need to be replaced. The actual
execute_pipeline() function for TellSeq Workflow() needs to be pasted
from StandardMetagenomicWorkflow() and modified to use the new
TellReadJob() class in mg-scripts.
